refactor: created components index exports files
