+++ components/button-link.tsx
@@ -6,7 +6,7 @@ interface Props {
   text: string
 }
 
-const ButtonLink: FC<Props> = ({ href, text }) => (
+export const ButtonLink: FC<Props> = ({ href, text }) => (
   <Link
     className="bg-primary-muted rounded-3xl font-bold text-primary-light px-3 py-2"
     href={href}
@@ -14,5 +14,3 @@ const ButtonLink: FC<Props> = ({ href, text }) => (
     {text}
   </Link>
 )
-
-export default ButtonLink
