+++ components/header/header.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link'
-import Navigation from '@/components/header/navigation'
+import { Navigation } from '@/components'
 
-const Header = () => (
+export const Header = () => (
   <header className="bg-primary-muted py-5 px-10 flex justify-between items-center sticky top-0 left-0 z-10">
     <div className="flex items-center gap-x-3">
       <h1 className="w-max border-2 text-primary-light font-bold py-1 px-2">
@@ -14,5 +14,3 @@ const Header = () => (
     <Navigation />
   </header>
 )
-
-export default Header
