+++ components/index.ts
@@ -1,0 +1,4 @@
+export * from './header'
+export * from './map'
+export * from './button-link'
+export * from './loader'
