+++ components/map/index.ts
@@ -1,0 +1,2 @@
+export * from './field'
+export * from './map-template'
