+++ components/map/map-template.tsx
@@ -1,11 +1,11 @@
 'use client'
 
-import Loader from '@/components/loader'
+import { Loader } from '@/components'
 import { useJsApiLoader } from '@react-google-maps/api'
 import { mapLibraries, mapOptions } from '@/constants'
 import { useEffect, useRef, useState } from 'react'
 import Field from '@/components/map/field'
-import { buildMapPlaceCard } from '@/utils'
+import { buildMapPlaceCard } from '@/app/map/utils'
 
 export const MapTemplate = () => {
   // Local states
